requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-filter==24.2
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
